requirements.txt
django
djangorestframework 
PyJWT
numpy==1.26.4
fasttext-wheel